pyproject.toml
[tool.poetry]
name = "sample-poetry"
version = "0.1.0"
description = ""
authors = ["sattosan <sato.yuya1211@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
autopep8 = "^1.6.0"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
